requirements.txt
streamlit
pandas
numpy
matplotlib
psycopg2-binary
openpyxl
scikit-learn